requirements.txt
absl-py==0.1.13
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
cycler==0.10.0
dicom==0.9.9.post1
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
med2image==1.1.1
nibabel==2.2.1
numpy==1.14.2
Pillow==5.1.0
protobuf==3.5.2.post1
pydicom==1.0.2
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
tqdm==4.23.0
Werkzeug==0.14.1
